ARM: dts: sun9i: a80-optimus: Enable hardware reset and HS-DDR for eMMC
mmc2 has a special pin for eMMC hardware reset, which is controllable
from the controller. Add the "mmc-cap-hw-reset" property to denote that
this controller supports this function, and the pins are actually used.

Also add the vqmmc regulator for signal voltage control, and increase
the signal drive strength for mmc2 pins, for HS-DDR mode support.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>